+++ src/arclet/alconna/analysis/special.py
@@ -1,4 +1,4 @@
-from ..components.output import output_send
+from ..components.output import output_manager
 from ..base import ShortcutOption
 from .parts import analyse_option
 from .analyser import Analyser
@@ -6,13 +6,13 @@ from .analyser import Analyser
 
 def handle_help(analyser: Analyser):
     analyser.current_index, analyser.content_index = analyser.head_pos
-    _help_param = [str(i) for i in analyser.rest_data() if i not in {"-h", "--help"}]
+    _help_param = [str(i) for i in analyser.release() if i not in {"-h", "--help"}]
 
     def _get_help():
         formatter = analyser.alconna.formatter_type(analyser.alconna)
         return formatter.format_node(_help_param)
 
-    output_send(analyser.alconna.name, _get_help).handle({}, is_raise_exception=analyser.is_raise_exception)
+    output_manager.get(analyser.alconna.name, _get_help).handle(is_raise_exception=analyser.is_raise_exception)
     return analyser.export(fail=True)
 
 
@@ -23,7 +23,7 @@ def handle_shortcut(analyser: Analyser):
             opt_v['name'], None if opt_v['command'] == "_" else analyser.converter(opt_v['command']),
             bool(opt_v.get('delete')), opt_v['expiration']
         )
-        output_send(analyser.alconna.name, lambda: msg).handle({}, is_raise_exception=analyser.is_raise_exception)
+        output_manager.get(analyser.alconna.name, lambda: msg).handle(is_raise_exception=analyser.is_raise_exception)
     except Exception as e:
-        output_send(analyser.alconna.name, lambda: str(e)).handle({}, is_raise_exception=analyser.is_raise_exception)
+        output_manager.get(analyser.alconna.name, lambda: str(e)).handle(is_raise_exception=analyser.is_raise_exception)
     return analyser.export(fail=True)